+++ app/javascript/controllers/service_controller.js
@@ -2,21 +2,9 @@ import { Controller } from "stimulus"
 
 
 export default class extends Controller {
-  static targets = ["connectedUrl", "serviceType", "contactEmails"]
-
-  connect(){}
+  static targets = ["contactEmails"]
 
   initialize(){
-    this.showConnectedUrl()
-  }
-
-  showConnectedUrl(event){
-    const serviceType = this.serviceTypeTarget.value
-    if (serviceType == "external" || serviceType == "open_access"){
-      this.connectedUrlTarget.classList.remove("hidden-fields");
-    } else {
-      this.connectedUrlTarget.classList.add("hidden-fields");
-    }
   }
 
   addNewEmailField(event) {